Fxied compilation warnings

+++ Flare/src/Flare/Renderer2D/Renderer2D.cpp
@@ -418,7 +418,7 @@ namespace Flare
 			else if (text[charIndex] == '\n')
 			{
 				position.x = 0;
-				position.y -= fontScale * metrics.lineHeight + lineHeightOffset;
+				position.y -= fontScale * (float)metrics.lineHeight + lineHeightOffset;
 				continue;
 			}
 			else if (!std::isspace(text[charIndex]))
